chat-backend/requirements.txt
fastapi[standard]
openai
pydantic_settings
pyyaml